Split main FastAPI app into routers
