+++ app/routes/__init__.py
@@ -1,0 +1,2 @@
+from routes.temperatures import router as temperatures_router
+from routes.cities import router as city_router
